Move asset certification header code to asset_util
This moves code from `http.rs` into the new `asset_util` crate.

This PR is step **2** of 4:
1. move code from `assets.rs` to the crate
2. move certification header code from `http.rs` to the crate
3. refactor the crate a bit to make the public interface nicer
4. use the crate in the other canisters
